requirements.txt
Flask==3.0.3
pandas==2.2.3
Pillow==11.1.0
python-docx==1.2.0
requests==2.32.2
